LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/path-objects/2d.path.isPointInPath.basic.1.worker-expected.txt


PASS isPointInPath() detects whether the point is inside the path 
